EN-4321 : Moved all status handler metrics from main.go in a another files. Fixed problem with maximum number of requests to github
