+++ cmd/node/metrics/machineStatistics.go
@@ -1,0 +1,88 @@
+package metrics
+
+import (
+	"errors"
+
+	"github.com/ElrondNetwork/elrond-go/core"
+	"github.com/ElrondNetwork/elrond-go/core/appStatusPolling"
+	"github.com/ElrondNetwork/elrond-go/core/statistics/machine"
+)
+
+// StartMachineStatisticsPolling will start read information about current  running machini
+func StartMachineStatisticsPolling(ash core.AppStatusHandler, pollingInterval int) error {
+	if ash == nil {
+		return errors.New("nil AppStatusHandler")
+	}
+
+	appStatusPollingHandler, err := appStatusPolling.NewAppStatusPolling(ash, pollingInterval)
+	if err != nil {
+		return errors.New("cannot init AppStatusPolling")
+	}
+
+	err = registerCpuStatistics(appStatusPollingHandler)
+	if err != nil {
+		return err
+	}
+
+	err = registerMemStatistics(appStatusPollingHandler)
+	if err != nil {
+		return err
+	}
+
+	err = registeNetStatistics(appStatusPollingHandler)
+	if err != nil {
+		return err
+	}
+
+	appStatusPollingHandler.Poll()
+
+	return nil
+}
+
+func registerMemStatistics(appStatusPollingHandler *appStatusPolling.AppStatusPolling) error {
+	memStats := &machine.MemStatistics{}
+	go func() {
+		for {
+			memStats.ComputeStatistics()
+		}
+	}()
+
+	return appStatusPollingHandler.RegisterPollingFunc(func(appStatusHandler core.AppStatusHandler) {
+		appStatusHandler.SetUInt64Value(core.MetricMemLoadPercent, memStats.MemPercentUsage())
+		appStatusHandler.SetUInt64Value(core.MetricMemTotal, memStats.TotalMemory())
+		appStatusHandler.SetUInt64Value(core.MetricMemUsedGolang, memStats.MemoryUsedByGolang())
+		appStatusHandler.SetUInt64Value(core.MetricMemUsedSystem, memStats.MemoryUsedBySystem())
+	})
+}
+
+func registeNetStatistics(appStatusPollingHandler *appStatusPolling.AppStatusPolling) error {
+	netStats := &machine.NetStatistics{}
+	go func() {
+		for {
+			netStats.ComputeStatistics()
+		}
+	}()
+
+	return appStatusPollingHandler.RegisterPollingFunc(func(appStatusHandler core.AppStatusHandler) {
+		appStatusHandler.SetUInt64Value(core.MetricNetworkRecvBps, netStats.BpsRecv())
+		appStatusHandler.SetUInt64Value(core.MetricNetworkRecvBpsPeak, netStats.BpsRecvPeak())
+		appStatusHandler.SetUInt64Value(core.MetricNetworkRecvPercent, netStats.PercentRecv())
+
+		appStatusHandler.SetUInt64Value(core.MetricNetworkSentBps, netStats.BpsSent())
+		appStatusHandler.SetUInt64Value(core.MetricNetworkSentBpsPeak, netStats.BpsSentPeak())
+		appStatusHandler.SetUInt64Value(core.MetricNetworkSentPercent, netStats.PercentSent())
+	})
+}
+
+func registerCpuStatistics(appStatusPollingHandler *appStatusPolling.AppStatusPolling) error {
+	cpuStats := &machine.CpuStatistics{}
+	go func() {
+		for {
+			cpuStats.ComputeStatistics()
+		}
+	}()
+
+	return appStatusPollingHandler.RegisterPollingFunc(func(appStatusHandler core.AppStatusHandler) {
+		appStatusHandler.SetUInt64Value(core.MetricCpuLoadPercent, cpuStats.CpuPercentUsage())
+	})
+}
